azure-ocr-trunk/requirements.txt
azure-ai-documentintelligence==1.0.1
azure-core==1.32.0
# azure-storage-blob==12.24.1
# llama_parse==0.6.9
openai==1.72.0
# pandas==2.2.3
python-dotenv==1.1.0
Requests==2.32.3
numpy==2.2.4
uvicorn==0.34.1
fastapi==0.115.12
redis==6.1.0
# sentence-transformers==4.1.0
# transformers==4.51.3S
python-dateutil==2.9.0.post0
scipy==1.15.2
